requirements.txt
numpy
pillow
mlx
h5py
